LayoutTests/http/tests/security/document-all-expected.txt

CONSOLE MESSAGE: line 1: TypeError: Illegal constructor
